Refactor shared tool contract test support
